+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/config.js
@@ -25,6 +25,36 @@ IpsuiteNavigatorConfig = Backbone.Model.extend({
         });
 
     },
-});
 
-ipsuiteNavigatorConfig = new IpsuiteNavigatorConfig();
+    // send current state of config to browser history.pushState
+    history_pushstate: function() {
+
+        // get current parameters from url
+        var url = $.url(window.location.href);
+        var params = url.param();
+
+        // aggregate parameters comprising viewer state, currently a 4-tuple
+        var state = {
+            mode: this.get('mode'),
+            context: this.get('context'),
+            project: this.get('project'),
+            datasource: this.get('datasource'),
+        };
+
+        // merge current viewer state
+        _(params).extend(state);
+
+        // clear empty parameters
+        for (key in params) {
+            var value = params[key];
+            if (_.isEmpty(value)) {
+                delete params[key];
+            }
+        }
+
+        // push parameters to browser history, changing the url in the location bar
+        history.pushState({id: 'url-clean'}, '', '?' + $.param(params));
+
+    },
+
+});